pyproject.toml
# This configuration idealy should be in backend/
# But Pycharm doesn't correctly recognize it

[tool.poetry]
name = "landsat-web-tracker"
version = "0.0.1"
description = ""
authors = ["EKK Ingarden"]
license = "GPL-3.0"
readme = "README.md"
packages = [{include = "backend"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.0"}
uvicorn = { extras = ["standard"], version = "^0.30.6" }
structlog = "^24.4.0"
sqlalchemy = "^2.0.35"
psycopg2 = "^2.9.9"
pydantic-settings = "^2.5.2"
pystac-client = "^0.8.3"
supabase = "^2.7.4"
planetary-computer = "^1.0.0"
httpx = "^0.27.2"
numpy = "^2.1.1"
pillow = "^10.4.0"

[tool.poetry.dev-dependencies]
pytest = "^8.3.3"

[tool.poetry.scripts]
openapi = "backend.utils.utils:write_openapi"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.9"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff]
line-length = 120
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001",
]

src = ["backend"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
